requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
python-dotenv
pydantic
sse-starlette